+++ tests/e2e/initialization/config.go
@@ -481,16 +481,17 @@ func updateTWAPGenesis(appGenState map[string]json.RawMessage) func(twapGenState
 				}
 
 				twapRecord := twaptypes.TwapRecord{
-					PoolId:                      balancerPool.Id,
-					Asset0Denom:                 denomPair.Denom0,
-					Asset1Denom:                 denomPair.Denom0,
-					Height:                      1,
-					Time:                        time.Date(2023, 0o2, 1, 0, 0, 0, 0, time.UTC), // some time in the past.
-					P0LastSpotPrice:             sp0,
-					P1LastSpotPrice:             sp1,
-					P0ArithmeticTwapAccumulator: osmomath.ZeroDec(),
-					P1ArithmeticTwapAccumulator: osmomath.ZeroDec(),
-					GeometricTwapAccumulator:    osmomath.ZeroDec(),
+					PoolId:      balancerPool.Id,
+					Asset0Denom: denomPair.Denom0,
+					Asset1Denom: denomPair.Denom0,
+					Height:      1,
+					Time:        time.Date(2023, 0o2, 1, 0, 0, 0, 0, time.UTC), // some time in the past.
+					// Note: truncation is acceptable as x/twap is guaranteed to work only on pools with spot prices > 10^-18.
+					P0LastSpotPrice:             sp0.Dec(),
+					P1LastSpotPrice:             sp1.Dec(),
+					P0ArithmeticTwapAccumulator: sdk.ZeroDec(),
+					P1ArithmeticTwapAccumulator: sdk.ZeroDec(),
+					GeometricTwapAccumulator:    sdk.ZeroDec(),
 					LastErrorTime:               time.Time{}, // no previous error
 				}
 				twapGenState.Twaps = append(twapGenState.Twaps, twapRecord)